Serialize into the buffer from the end backwards
This will allow two tables to refer to a common child
in the buffer without having to do a topological sort
of the whole object graph before writing anything.

Note. This commit doesn't actually cause a common child
to only be serialized once, that's coming in a future commit.

The big change from the user perspective is that the
`Write` methods now return an offset to the start of
the data within the buffer, not the end of the data.